cSRF refactoring
