+++ calvincTools/utils/forms/definitions/cQFormFieldDef.py
@@ -5,7 +5,6 @@ from enum import Enum
 from PySide6.QtCore import Qt
 from PySide6.QtWidgets import QWidget
 
-from .cQFmConstants import cQFmConstants
 
 @dataclass(frozen=True)
 class cQFormFieldDef:
