+++ calvincTools/utils/forms/forms/cSRFSingleRecordForm.py
@@ -1,0 +1,628 @@
+from calvincTools.utils.cQWidgets import cGridWidget, cstdTabWidget
+from calvincTools.utils.forms.definitions.cQFormBtnDef import cQFormBtnDef
+from calvincTools.utils.forms.definitions.cQFormFieldDef import cQFormFieldDef
+from calvincTools.utils.forms.definitions.cQFormLayout import cQFormLayout
+from calvincTools.utils.forms.forms.cSRF_FormUI_Base import cSRF_FormUI_Base
+from calvincTools.utils.forms.forms.cSRF_Formdb_Base import cSRF_Formdb_Base
+from calvincTools.utils.forms.widgets.cQFmNameLabel import cQFmNameLabel
+from calvincTools.utils.forms.widgets.cSimpRecFmElement_Base import cSimpRecFmElement_Base
+from calvincTools.utils.messageBoxes import areYouSure
+
+
+import qtawesome
+from PySide6.QtCore import Slot
+from PySide6.QtGui import QFont
+from PySide6.QtWidgets import QGridLayout, QHBoxLayout, QLabel, QMessageBox, QPushButton, QStatusBar, QVBoxLayout, QWidget
+from sqlalchemy import func
+from sqlalchemy.orm import Session, sessionmaker
+
+
+from typing import Any, List, Type
+
+
+class cSRFSingleRecordForm(cSRF_FormUI_Base, cSRF_Formdb_Base):
+    """
+    Base class for single record forms. Inherits from both cSRF_FormUI_Base and cSRF_Formdb_Base to combine UI and db functionality.
+
+    children must implement:
+    Define form fields - to be implemented by subclass
+
+    def defineFields(self):
+        Define the form fields.
+
+        This method should be implemented by subclasses to define the form fields
+        and their properties. It should populate self._field_defs with a list of cQFormFieldDef instances.
+
+    """
+    def __init__(self,
+        formname: str|None = None,
+        field_defs: List[cQFormFieldDef] | None = None,
+        model: Type[Any]|None = None,
+        ssnmaker: sessionmaker[Session] | None = None,
+        parent: QWidget | None = None,
+        *args, **kwargs):
+
+        self._formname = getattr(self, '_formname', None)
+        if not self._formname:
+            self._formname = formname if formname else 'Form'
+
+        super().__init__(field_defs=field_defs, model=model, ssnmaker=ssnmaker, parent=parent, *args, **kwargs)
+    # __init__
+
+    ######################################################
+    ########    Layout construction
+
+    def _buildFormLayout(self) -> cQFormLayout:
+        """Build the form layout for cSimpleRecordForm.
+
+        Returns:
+            QFormLayout instance
+        """
+
+        layoutMain = QVBoxLayout(self)
+        layoutFormHdr = QHBoxLayout()
+        layoutForm = cGridWidget(scrollable=True)
+        layoutFormFixedTop = QGridLayout()
+        layoutFormPages = cstdTabWidget()
+        layoutFormFixedBottom = QGridLayout()
+        layoutButtons = QVBoxLayout()  # may get redefined in _addActionButtons
+        statusBar = QStatusBar(self)
+
+        # should this be in _finalizeMainLayout instead?
+        layoutForm.addLayout(layoutFormFixedTop, 0, 0)
+        layoutForm.addWidget(layoutFormPages, 1, 0)
+        layoutForm.addLayout(layoutFormFixedBottom, 2, 0)
+
+        assert isinstance(self._formname, str), "_formname must be set before building form layout"
+        lblFormName = cQFmNameLabel(self.tr(self._formname), self)
+        layoutFormHdr.addWidget(lblFormName)
+
+        newrecFlag = QLabel("New Record", self)
+        fontNewRec = QFont()
+        fontNewRec.setBold(True)
+        fontNewRec.setPointSize(10)
+        fontNewRec.setItalic(True)
+        newrecFlag.setFont(fontNewRec)
+        newrecFlag.setStyleSheet("color: red;")
+        layoutFormHdr.addWidget(newrecFlag)
+
+        # put it together
+        layoutMain.addLayout(layoutFormHdr)
+        layoutMain.addWidget(layoutForm)
+        layoutMain.addLayout(layoutButtons)
+        layoutMain.addWidget(statusBar)
+
+        self.setWindowTitle(self.tr(self._formname))
+
+        rtnval = cQFormLayout(
+            main=layoutMain,
+            header=layoutFormHdr,
+            form=layoutForm,
+            fixed_top=layoutFormFixedTop,
+            pages=layoutFormPages,
+            fixed_bottom=layoutFormFixedBottom,
+            buttons=layoutButtons,
+            status_bar=statusBar,
+
+            lblFormName = lblFormName,
+            newrecFlag = newrecFlag,
+            )
+
+        return rtnval
+    # _buildFormLayout
+
+    ######################################################
+    ########    field and Widget placement
+
+    def defineActionButtons(self):
+        _iconlib = qtawesome.icon
+        r = [
+            cQFormBtnDef(text="First", icon=_iconlib("mdi.page-first"), action=self.on_loadfirst_clicked),
+            cQFormBtnDef(text="Previous", icon= _iconlib("mdi.arrow-left-bold"), action=self.on_loadprev_clicked),
+            cQFormBtnDef(text="Next", icon=_iconlib("mdi.arrow-right-bold"), action=self.on_loadnext_clicked),
+            cQFormBtnDef(text="Last", icon=_iconlib("mdi.page-last"), action=self.on_loadlast_clicked),
+            cQFormBtnDef(type=cQFormBtnDef.ButtonType.NEW_HSECTION),
+            cQFormBtnDef(text="Add", icon=_iconlib("mdi.plus"), action=self.on_add_clicked),
+            cQFormBtnDef(text="Save", icon=_iconlib("mdi.content-save"), commitBtn=True, action=self.on_save_clicked),
+            cQFormBtnDef(text="Delete", icon=_iconlib("mdi.delete"), action=self.on_delete_clicked),
+            cQFormBtnDef(text="Cancel", icon=_iconlib("mdi.cancel"), action=self.on_cancel_clicked),
+            ]
+        return r
+
+    def _addActionButtons(self,
+            ActionButtons:List[cQFormBtnDef]|None = None,
+            ) -> None:
+        """Add action buttons to the form.
+        """
+
+        Actns = ActionButtons if ActionButtons is not None else self.defineActionButtons()
+        if Actns is None:
+            return
+
+        layoutButtons = self._layouts.buttons
+
+        innerLayout = QHBoxLayout()
+
+        for btndef in Actns:
+            if btndef.type == cQFormBtnDef.ButtonType.NEW_VSECTION:
+                layoutButtons.addLayout(innerLayout)
+                innerLayout = QHBoxLayout()
+            elif btndef.type == cQFormBtnDef.ButtonType.NEW_HSECTION:
+                innerLayout.addSpacing(20)
+            elif btndef.type != cQFormBtnDef.ButtonType.NORMAL:
+                raise ValueError(f"unknown button type {btndef.type}")
+            else:
+                btn = QPushButton(btndef.text)
+                if btndef.icon is not None:
+                    btn.setIcon(btndef.icon)
+                if callable(btndef.action):
+                    btn.clicked.connect(btndef.action)
+                if btndef.commitBtn:
+                    self.btnCommit = btn
+                innerLayout.addWidget(btn)
+            # endif button type
+        # endfor btndef om Actns
+        layoutButtons.addLayout(innerLayout)
+    # _addNavButtons
+
+    ######################################################
+    ########    Display 
+
+    def initialdisplay(self):
+        """Initialize and display the first record.
+
+        Initializes a new record and loads the first record from the database.
+        """
+        self.initializeRec()
+        self.on_loadfirst_clicked()
+    # initialdisplay()
+
+    def fillFormFromcurrRec(self):
+        """Load the current record into all form fields.
+
+        Updates all field widgets with values from the current record
+        and updates the dirty and new record flags.
+        """
+        for widgStructure in self._formWidgets.values():
+            widg = widgStructure.widget
+            if isinstance(widg, cSimpRecFmElement_Base):
+                widg.loadFromRecord(self.currRec())
+
+        self.showNewRecordFlag()
+        self.showCommitButton()
+        # self.setDirty(False) - nope, don't need to set form dirty state here - isDirty checks individual fields
+    # fillFormFromRec
+
+    # TODO: wrap with fillFormFromcurrRec
+    # TODO: play with positioning of new record flag
+    def showNewRecordFlag(self) -> None:
+        """Show or hide the 'New Record' flag based on current record state."""
+        nrf = getattr(self, '_newrecFlag', None)
+        if not isinstance(nrf, QWidget):
+            return
+        nrf.setVisible(self.isNewRecord())
+
+    def showCommitButton(self) -> None:
+        """Show the commit button if the record is dirty."""
+        btnCommit = getattr(self, 'btnCommit', None)
+        if not isinstance(btnCommit, QWidget):
+            return
+        btnCommit.setEnabled(self.isDirty())
+    # showCommitButton
+
+    ##################################################
+    ########    Record Navigation 
+
+    def isit_OKToLeaveRecord(self) -> bool:
+        """
+        Check if the form is dirty. If so, prompt user.
+        Returns True if it is safe to proceed with navigation, False otherwise.
+        """
+        if not self.isDirty():
+            return True
+
+        choice = areYouSure(
+            self,
+            "Unsaved changes",
+            "You have unsaved changes. Save before continuing?",
+            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No | QMessageBox.StandardButton.Cancel,
+            QMessageBox.StandardButton.Yes,
+            )
+
+        if choice == QMessageBox.StandardButton.Yes:
+            self.on_save_clicked()
+            return True
+
+        elif choice == QMessageBox.StandardButton.No:
+            # Discard changes -> reload current record fresh
+            if self.currRec():
+                self.fillFormFromcurrRec()
+            return True
+
+        else:  # Cancel
+            return False
+    # isit_OKToLeaveRecord
+
+    def _navigate_to(self, rec_id: int):
+        """Navigate safely to a record (with save/discard prompt if dirty)."""
+        if not self.isit_OKToLeaveRecord():
+            return  # Cancel pressed → stay put
+
+        self._load_record_by_id(rec_id)
+    # _navigate_to
+
+    def get_prev_record_id(self, recID:int) -> int:
+        """Get the ID of the previous record.
+
+        Args:
+            recID (int): Current record ID.
+
+        Returns:
+            int: ID of the previous record, or None if no previous record exists.
+        """
+        ssnmkr = self.ssnmaker()
+        assert ssnmkr is not None, "Sessionmaker must be set before touching the database"
+        pKey = self.primary_key()
+        with ssnmkr() as session:
+            prev_id = session.query(func.max(pKey)).where(pKey < recID).scalar()
+        return prev_id
+    def get_next_record_id(self, recID:int) -> int:
+        """Get the ID of the next record.
+
+        Args:
+            recID (int): Current record ID.
+
+        Returns:
+            int: ID of the next record, or None if no next record exists.
+        """
+        ssnmkr = self.ssnmaker()
+        pKey = self.primary_key()
+        assert ssnmkr is not None, "Sessionmaker must be set before touching the database"
+        with ssnmkr() as session:
+            next_id = session.query(func.min(pKey)).where(pKey > recID).scalar()
+        return next_id
+
+    def on_loadfirst_clicked(self):
+        """Load the first record in the database."""
+        # determine minimum id in database and load it
+        ssnmkr = self.ssnmaker()
+        pKey = self.primary_key()
+        assert ssnmkr is not None, "Sessionmaker must be set before touching the database"
+        with ssnmkr() as session:
+            min_id = session.query(func.min(pKey)).scalar()
+            if min_id:
+                self._navigate_to(min_id)
+
+    def on_loadprev_clicked(self):
+        """Load the previous record in the database."""
+        # determine previous id in database and load it
+        currRec = self.currRec()
+        pKey = self.primary_key()
+        currID = getattr(currRec, pKey.key)
+        prev_id = self.get_prev_record_id(currID)
+        if prev_id:
+            self._navigate_to(prev_id)
+
+    def on_loadnext_clicked(self):
+        """Load the next record in the database."""
+        # determine next id in database and load it
+        currRec = self.currRec()
+        pKey = self.primary_key()
+        currID = getattr(currRec, pKey.key)
+        next_id = self.get_next_record_id(currID)
+        if next_id:
+            self._navigate_to(next_id)
+
+    def on_loadlast_clicked(self):
+        """Load the last record in the database."""
+        # determine maximum id in database and load it
+        ssnmkr = self.ssnmaker()
+        pKey = self.primary_key()
+        assert ssnmkr is not None, "Sessionmaker must be set before touching the database"
+        with ssnmkr() as session:
+            max_id = session.query(func.max(pKey)).scalar()
+            if max_id:
+                self._navigate_to(max_id)
+
+    ##########################################
+    ########    Create
+
+    def initializeRec(self, initializeTo=None):
+        """
+        Initialize a new record with default values.
+
+        implementation should call fillFormFromcurrRec() after setting default values in self.currRec
+        """
+        modlType = self.ORMmodel() if initializeTo is None else initializeTo
+        assert modlType is not None, "ORMmodel must be set before initializing record"
+        self.setcurrRec(modlType())
+    # initializeRec
+
+    def on_add_clicked(self):
+        """
+        Add a new record to the database: initialize, set defaults and save.
+        No, don't save. reserve that for the save button.
+        """
+        # if dirty, ask to save
+        if not self.isit_OKToLeaveRecord():
+            return
+
+        self.initializeRec()
+        self.fillFormFromcurrRec()
+    # add_record
+
+    ##########################################
+    ########    Read
+
+    # # --- Lookup navigation ---
+    def _load_record_by_id(self, pk_val):
+        """Low-level load (assumes it's safe to replace current record)."""
+        ssnmkr = self.ssnmaker()
+        assert ssnmkr is not None, "Sessionmaker must be set before touching the database"
+        with ssnmkr() as session:
+            modl = self.ORMmodel()
+            assert modl is not None, "ORMmodel must be set before loading record"
+            rec = session.get(modl, pk_val)
+            if rec is None:
+                self.showError(f"No Record with id {pk_val}")
+                return
+            else:
+                # detach rec from session and make it the current record
+                session.expunge(rec)
+                self.setcurrRec(rec)
+                self.fillFormFromcurrRec()
+            # endif rec 
+        #endwith session
+    # load_record_by_id
+
+    def load_record(self, recindex: int):
+        """
+        Load a record from the database.
+        NOTE: For this class, recindex is the id of the record to load, not the index.
+
+        Args:
+            recindex (int): The ID of the record to load.
+        """
+        self._load_record_by_id(recindex)
+    # load_record
+
+    def load_record_by_field(self, field: str | Any, value: Any) -> None:
+        """
+        field may be either:
+          - a string (column name), or
+          - an ORM field object (MyModel.name).
+        """
+        if not self.isit_OKToLeaveRecord():
+            return  # Cancel pressed → stay put
+
+        if isinstance(field, str):
+            orm_field = getattr(self._ORMmodel, field)
+        else:
+            orm_field = field
+
+        ssnmkr = self.ssnmaker()
+        assert ssnmkr is not None, "Sessionmaker must be set before touching the database"
+        with ssnmkr() as session:
+            modl = self.ORMmodel()
+            assert modl is not None, "ORMmodel must be set before loading record"
+            rec = session.query(modl).filter(orm_field == value).first()
+            if rec is None:
+                self.showError(f"No Record with {orm_field.key} == {value}")
+                return
+            else:
+                # detach rec from session and make it the current record
+                session.expunge(rec)
+                self.setcurrRec(rec)
+                self.fillFormFromcurrRec()
+            # endif rec 
+        #endwith session
+    # load_record_by_field
+
+    @Slot()
+    def lookup_and_load(self, fld: str, value: Any):
+        """Load a record by looking up a field value.
+
+        Args:
+            fld (str): Field name to search.
+            value (Any): Value to search for (can be dict with 'text' key, or plain value).
+        """
+        value = value.get('text', value) if isinstance(value, dict) else (getattr(value, 'text', value) if hasattr(value, 'text') else value)
+        self.load_record_by_field(fld, value)
+    # lookup_CIMSNum
+
+    ##########################################
+    ########    Update
+
+    @Slot()
+    def _on_field_changed(self, widget, defn: cQFormFieldDef):
+        value = widget.Value()
+
+        if defn.transform:
+            value = defn.transform(value)
+
+        if defn.on_change:
+            defn.on_change(value)
+
+        self.showCommitButton()
+    # _on_field_changed
+
+    @Slot()
+    def on_save_clicked(self, *_):
+        """
+        Collects field values from adapters/subforms, writes them into currRec,
+        and persists via a short-lived session.
+        """
+        currRec = self.currRec()
+        if not currRec:
+            return
+
+        try:
+            # Push data from form -> ORM object, except for subforms - they must come after the main record is saved
+            # for fldName, fldDef in self._field_defs_by_name.items():
+            for fldName, fldStruct in self._formWidgets.items():
+                fldDef = fldStruct.defn
+                widget = fldStruct.widget
+                isSubFormElmnt = fldDef.field_type == cQFormFieldDef.cQFormFieldType.SUBFORM
+                if not isSubFormElmnt:      # subforms handled after main record is saved
+                    if isinstance(widget, cSimpRecFmElement_Base):
+                        widget.saveToRecord(currRec)
+            # endfor fldDef in self.fieldDefs
+
+            # Persist using a short-lived session
+            ssnmkr = self.ssnmaker()
+            assert ssnmkr is not None, "Sessionmaker must be set before touching the database"
+            with ssnmkr(expire_on_commit=False) as session:
+                merged = session.merge(currRec)
+                session.flush()
+                session.refresh(merged)
+
+                pKey = self.primary_key()
+                recID = getattr(merged, pKey.key)   # no change for existing record; loads new id for a new one
+                # should I copy other fields, too?
+
+                # session.expunge(self.currRec) # not needed, since self.currRec not bound to session
+
+                session.commit()
+            # endwith session
+
+            # now handle subforms
+            # for fldName, fldDef in self._field_defs_by_name.items():
+            for fldName, fldStruct in self._formWidgets.items():
+                fldDef = fldStruct.defn
+                widget = fldStruct.widget
+                isSubFormElmnt = fldDef.field_type == cQFormFieldDef.cQFormFieldType.SUBFORM
+                if isSubFormElmnt:
+                    if isinstance(widget, cSimpRecFmElement_Base):
+                        widget.saveToRecord(currRec)
+            # endfor fldDef in self.fieldDefs
+
+            # reload the record (repaints screen, gets db defaults and new id, if any)
+            self.repopLookups()
+            self._load_record_by_id(recID)
+
+            # all this not needed because of reload
+            # # Reset dirty flags (both form and adapters)
+            # for fldName, fldDef in self.fieldDefs.items():
+            #     widget = fldDef.get("widget")
+            #     if widget:
+            #         widget.setDirty(False)
+            # self.setDirty(self, False)
+
+            # # Clear new record flag
+            # assert not self.isNewRecord()
+            # self.showNewRecordFlag(False)
+
+        except Exception as e:
+            self.showError(str(e), "Error saving record")
+    # on_save_clicked
+
+    ##########################################
+    ########    Delete
+
+    # TODO: confirm delete
+    @Slot()
+    def on_delete_clicked(self):
+        """Handle delete button click.
+
+        Prompts for confirmation, then deletes the current record and navigates
+        to a neighboring record.
+        """
+        currRec = self.currRec()
+        if not currRec:
+            return
+
+        pKey = self.primary_key()
+        keyID = getattr(currRec, pKey.key)
+
+        if not self.isit_OKToLeaveRecord():
+            return  # Cancel pressed → stay put
+
+        confirm = areYouSure(
+            self,
+            "Delete record",
+            f"Really delete record {keyID}?",
+            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
+            QMessageBox.StandardButton.No,
+        )
+        if confirm != QMessageBox.StandardButton.Yes:
+            return
+
+        # Actually delete
+        ssnmkr = self.ssnmaker()
+        assert ssnmkr is not None, "Sessionmaker must be set before touching the database"
+        modl = self.ORMmodel()
+        assert modl is not None, "ORMmodel must be set before deleting record"
+        with ssnmkr() as session:
+            rec = session.get(modl, keyID)
+            if rec:
+                session.delete(rec)
+                session.commit()
+
+        self.repopLookups()
+        # Navigate to neighbor, or clear form if none
+        next_id = self.get_next_record_id(keyID)
+        prev_id = self.get_prev_record_id(keyID)
+        target_id = next_id or prev_id
+        if target_id:
+            self._load_record_by_id(target_id)
+        else:
+            self.initializeRec()
+            self.fillFormFromcurrRec()
+    # on_delete_clicked
+
+    # ##########################################
+    # ########    Record Status
+
+    def isNewRecord(self) -> bool:
+        """Check if the current record is a new (unsaved) record.
+
+        Returns:
+            bool: True if the current record has no primary key value, False otherwise.
+        """
+        pKey = self.primary_key()
+        currRec = self.currRec()
+        return currRec is None or getattr(currRec, pKey.key) is None
+
+    @Slot()
+    def setDirty(self, dirty: bool|None = None):
+        """Poll children for dirty state and update save button."""
+        # removed code to set dirty state of children, since each adapter handles its own dirty state internally and the form's dirty state is determined by polling the children when needed
+
+        # Enable save button if anything is dirty
+        btnCommit = getattr(self, 'btnCommit', None)
+        if hasattr(btnCommit, 'setEnabled'):
+            btnCommit.setEnabled(self.isDirty()) # type: ignore
+    # setFormDirty
+
+    def isDirty(self) -> bool:
+        """Check if any form element is dirty."""
+        # any() stops and returns True as soon as it finds the first True
+        return any(el.widget.isDirty() for el in self._formWidgets.values())       # type: ignore
+    # isDirty
+
+    def on_cancel_clicked(self):
+        """Handle the Cancel button click by closing the form.
+
+        Note:
+            Currently just closes the form without any confirmation.
+        """
+        #for now, just close form
+        self.close()
+    # cancel_record
+
+    def repopLookups(self) -> None:
+        """Refresh all lookup widgets with current database values."""
+        for lkupname, lkupwdgt in self._lookupFrmElements.items():
+                lkupwdgt.refreshChoices()   # type: ignore
+    # repopLookups
+
+    ###############################################
+    ###############################################
+    # TO BE ADDED LATER:
+    def beforeLoad(self, rec): pass
+    def afterLoad(self, rec): pass
+    def beforeSave(self): pass
+    def afterSave(self): pass
+# cSRFSingleRecordForm
+    def endofclass(self):
+        pass
